Week09/Day4/Exercise/requirements.txt
python-dotenv
streamlit
openai
mcp
